requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
apscheduler
openai
httpx
pymysql
pytz
motor